fix(pb): optimize actor key allocation
