+++ engine/packages/epoxy/tests/kv_get_optimistic.rs
@@ -31,6 +31,7 @@ async fn get_with_behavior(
 		replica_id,
 		key: key.to_vec(),
 		caching_behavior,
+		target_replicas: None,
 		save_empty: false,
 	})
 	.await
