+++ engine/packages/pegboard/src/ops/actor/get_reservation_for_key.rs
@@ -33,6 +33,7 @@ pub async fn pegboard_actor_get_reservation_for_key(
 			replica_id: ctx.config().epoxy_replica_id(),
 			key: keys::subspace().pack(&reservation_key),
 			caching_behavior: epoxy::protocol::CachingBehavior::Optimistic,
+			target_replicas: None,
 			save_empty: false,
 		})
 		.await?
